# backend/src/test/resources/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 교원사이트 공통기능 API
  version: 1.0.0
  description: CMS-1351 test0731 1차 완료 목표 HTTP API 계약
servers:
- url: /
paths:
...
  /api/organizations:
    get:
      operationId: listOrganizations
      summary: 조직 목록 조회
      x-related-requirements:
      - REQ-002
      - REQ-010
      - REQ-027
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 검증 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
      - sessionCookie: []
      parameters:
      - *id001
      - *id002
      - *id003
      - name: organizationCode
        in: query
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
      - name: filter
        in: query
        required: false
        schema:
          type: string
      x-roles:
      - authenticated
  /api/organizations/tree:
    get:
      operationId: getOrganizationTree
      summary: 조직 계층 조회
      x-related-requirements:
      - REQ-002
      - REQ-027
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 검증 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
      - sessionCookie: []
      parameters:
      - &id005
        $ref: '#/components/parameters/Page'
      - &id006
        $ref: '#/components/parameters/Size'
      - &id007
        $ref: '#/components/parameters/Sort'
      x-roles:
      - authenticated
  /api/organizations/{organizationCode}/relations:
    put:
      operationId: saveOrganizationRelation
      summary: 조직 관계 적용기간 저장
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-002
      - REQ-017
      - REQ-021
      - REQ-027
      x-business-rules:
      - effective_end_date는 effective_start_date보다 이전일 수 없다
      - 조직 기본정보는 KORUS Mock snapshot 원천을 직접 수정하지 않는다
      x-side-effects:
      - organization_relation row를 저장한다
      - change_history에 상위조직 변경 전후 값을 기록한다
      x-required-tests:
      - 'happy: 유효한 조직 관계 적용기간 저장 요청 -> 후속 조회에서 관련 table 변경값이 ApiResponse.data에 표시된다'
      - 'auth: 세션 없는 조직 관계 적용기간 저장 요청 -> HTTP 401과 ApiError.message가 반환된다'
      - 'validation: 필수 식별자 누락 -> ApiError.fields에 해당 field 오류가 포함된다'
      - 'business: effective_end_date는 effective_start_date보다 이전일 수 없다 위반 -> 대상 table의 기존 상태가 변경되지 않는다'
      - 'side-effect: 조직 관계 적용기간 저장 성공 -> change_history 또는 관련 상태 row가 생성된다'
      responses:
        '200':
          description: 저장 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 검증 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      parameters:
      - name: organizationCode
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      x-roles:
      - authenticated
...
  /api/menus:
    get:
      operationId: listMenus
      summary: 메뉴 실행정보 조회
      x-related-requirements:
      - REQ-007
      - REQ-027
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 검증 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
      - sessionCookie: []
      parameters: *id004
      x-roles:
      - authenticated
    post:
      operationId: createMenu
      summary: 메뉴 실행정보 등록
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-007
      - REQ-017
      - REQ-021
      - REQ-027
      x-business-rules:
      - menu_name과 screen_id 또는 url 연결정보가 필요하다
      - 사용 중인 메뉴 삭제 대신 비활성화를 우선한다
      x-side-effects:
      - menu row를 생성한다
      - change_history에 메뉴 등록 값을 기록한다
      x-required-tests:
      - 'happy: 유효한 메뉴 실행정보 등록 요청 -> 후속 조회에서 관련 table 변경값이 ApiResponse.data에 표시된다'
      - 'auth: 세션 없는 메뉴 실행정보 등록 요청 -> HTTP 401과 ApiError.message가 반환된다'
      - 'validation: 필수 식별자 누락 -> ApiError.fields에 해당 field 오류가 포함된다'
      - 'business: menu_name과 screen_id 또는 url 연결정보가 필요하다 위반 -> 대상 table의 기존 상태가 변경되지 않는다'
      - 'side-effect: 메뉴 실행정보 등록 성공 -> change_history 또는 관련 상태 row가 생성된다'
      responses:
        '200':
          description: 저장 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 검증 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      x-roles:
      - authenticated
  /api/menus/{menuId}:
    put:
      operationId: updateMenu
      summary: 메뉴 실행정보 수정
      security:
      - sessionCookie: []
      x-related-requirements:
      - REQ-007
      - REQ-017
      - REQ-019
      - REQ-021
      - REQ-027
      x-business-rules:
      - 존재하는 menu_id만 수정할 수 있다
      - 사용 중인 메뉴는 물리삭제보다 use_yn=N 비활성화를 우선한다
      x-side-effects:
      - menu 실행정보를 갱신한다
      - change_history에 메뉴 수정 전후 값을 기록한다
      x-required-tests:
      - 'happy: 유효한 메뉴 실행정보 수정 요청 -> 후속 조회에서 관련 table 변경값이 ApiResponse.data에 표시된다'
      - 'auth: 세션 없는 메뉴 실행정보 수정 요청 -> HTTP 401과 ApiError.message가 반환된다'
      - 'validation: 필수 식별자 누락 -> ApiError.fields에 해당 field 오류가 포함된다'
      - 'business: 존재하는 menu_id만 수정할 수 있다 위반 -> 대상 table의 기존 상태가 변경되지 않는다'
      - 'side-effect: 메뉴 실행정보 수정 성공 -> change_history 또는 관련 상태 row가 생성된다'
      responses:
        '200':
          description: 저장 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 검증 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      parameters:
      - name: menuId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      x-roles:
      - authenticated
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
  schemas:
    ApiError:
      type: object
      required:
      - message
      - fields
      properties:
        message:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
    ApiResponse:
      type: object
      required:
      - success
      properties:
        success:
          type: boolean
        data: {}
        error:
          $ref: '#/components/schemas/ApiError'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
...
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
    Sort:
      name: sort
      in: query
      schema:
        type: string
...